Update core functions to use new video service

+++ video_io/client.py
@@ -1,3 +1,5 @@
+import video_io.config
+import asyncio
 import concurrent.futures
 import json
 import os
@@ -5,6 +7,7 @@ from pathlib import Path
 import re
 import subprocess
 from typing import List, Dict
+import logging
 
 from auth0.v3.management.users import Users
 from auth0.v3.management.clients import Clients
@@ -15,10 +18,10 @@ import jmespath
 import requests
 from tenacity import retry, wait_random, stop_after_attempt
 
+logger = logging.getLogger(__name__)
 
 class SyncError(Exception):
     pass
-
 
 class RequestError(Exception):
 
@@ -26,29 +29,17 @@ class RequestError(Exception):
         super().__init__(f"unexpected api response - {response.status_code}")
         self.response = response
 
-
-
 class UnableToAuthenticate(Exception):
     pass
 
-
-
-AUTH0_DOMAIN = os.environ.get("AUTH0_DOMAIN")
-API_AUDIENCE = os.environ.get("API_AUDIENCE")
-CLIENT_ID = os.environ.get("AUTH0_CLIENT_ID")
-CLIENT_SECRET = os.environ.get("AUTH0_CLIENT_SECRET")
-VIDEO_STORAGE_LOCAL_CACHE_DIRECTORY = os.environ.get("VIDEO_STORAGE_LOCAL_CACHE_DIRECTORY", "/data")
-MAX_SYNC_WORKERS = os.environ.get("MAX_SYNC_WORKERS", 10)
-VIDEO_STORAGE_URL = os.environ.get("VIDEO_STORAGE_URL", "https://video.api.wildflower-tech.org")
-
-
 @ttl_cache(ttl=60 * 60 * 4)
-def client_token(audience=None):
-    get_token = GetToken(AUTH0_DOMAIN, timeout=10)
+def client_token(auth_domain, audience, client_id=None, client_secret=None):
+    get_token = GetToken(auth_domain, timeout=10)
     token = get_token.client_credentials(
-        CLIENT_ID,
-        CLIENT_SECRET,
-        audience if audience is not None else API_AUDIENCE)
+        client_id,
+        client_secret,
+        audience
+    )
     api_token = token['access_token']
     return api_token
 
@@ -66,22 +57,6 @@ def get_video_file_details(path):
         path,
     ], capture_output=True)
     return json.loads(ffprobe_out.stdout)
-
-
-def client_from_honeycomb_settings(client=None, token_uri=None, audience=None, client_id=None, client_secret=None):
-    if client is not None:
-        token = client.client.accessToken
-    elif client_id is not None and client_secret is not None:
-        get_token = GetToken(AUTH0_DOMAIN, timeout=10)
-        response = get_token.client_credentials(
-            client_id,
-            client_secret,
-            audience if audience is not None else API_AUDIENCE)
-        token = response['access_token']
-    else:
-        token = client_token(audience)
-    return VideoStorageClient(token)
-
 
 CACHE_PATH_FILE = re.compile('^(?P<environment_id>[a-fA-F0-9-]*)/(?P<camera_id>[a-fA-F0-9-]*)/(?P<year>[0-9]{4})/(?P<month>[0-9]{2})/(?P<day>[0-9]{2})/(?P<hour>[0-9]{2})/(?P<file>.*)$')
 CACHE_PATH_HOUR = re.compile('^(?P<environment_id>[a-fA-F0-9-]*)/(?P<camera_id>[a-fA-F0-9-]*)/(?P<year>[0-9]{4})/(?P<month>[0-9]{2})/(?P<day>[0-9]{2})/(?P<hour>[0-9]{2})$')
@@ -114,38 +89,62 @@ def chunks(lst, n):
     for i in range(0, len(lst), n):
         yield lst[i:i + n]
 
-
 class VideoStorageClient:
 
-    def __init__(self, token=None):
-        self.DOMAIN = VIDEO_STORAGE_URL
+    def __init__(
+        self,
+        token=None,
+        cache_directory=video_io.config.VIDEO_STORAGE_LOCAL_CACHE_DIRECTORY,
+        url=video_io.config.VIDEO_STORAGE_URL,
+        auth_domain=video_io.config.VIDEO_STORAGE_AUTH_DOMAIN,
+        audience=video_io.config.VIDEO_STORAGE_AUDIENCE,
+        client_id=video_io.config.VIDEO_STORAGE_CLIENT_ID,
+        client_secret=video_io.config.HONEYCOMB_CLIENT_SECRET
+    ):
+        self.CACHE_DIRECTORY = cache_directory
+        self.URL = url
         if token is not None:
             self.token = token
         else:
-            self.token = client_token()
+            self.token = client_token(
+                auth_domain=auth_domain,
+                audience=audience,
+                client_id=client_id,
+                client_secret=client_secret
+            )
 
     async def get_videos(self, environment_id, start_date, end_date, camera_id=None, destination=None):
-        prefix = destination if destination is not None else VIDEO_STORAGE_LOCAL_CACHE_DIRECTORY
-        os.makedirs(prefix, exist_ok=True)
+        if destination is None:
+            destination=self.CACHE_DIRECTORY
+        os.makedirs(destination, exist_ok=True)
         meta = self.get_videos_metadata_paginated(environment_id, start_date, end_date, camera_id)
-        async for vid_meta in meta:
-            await self.get_video(vid_meta["meta"]["path"], destination)
+        futures = []
+        with concurrent.futures.ThreadPoolExecutor(max_workers=4) as e:
+            async for vid_meta in meta:
+                f = e.submit(asyncio.run, self.get_video(path=vid_meta["meta"]["path"], destination=destination))
+                futures.append(f)
+        res = [r for r in concurrent.futures.as_completed(futures)]
 
     async def get_video(self, path, destination):
-        request = {
-            "method": "GET",
-            "url": f'{self.DOMAIN}/video/{path}',
-            "headers": {
-                "Authorization": f"bearer {self.token}",
-            },
-        }
-        response = requests.request(**request)
-        if response.status_code == 200:
-            p = Path(destination).joinpath(path)
-            pp = p.parent
-            if not pp.exists():
-                pp.mkdir(parents=True, exist_ok=True)
-            p.write_bytes(response.content)
+        p = Path(destination).joinpath(path)
+        if not p.is_file():
+            logger.info('Downloading video file %s', path)
+            request = {
+                "method": "GET",
+                "url": f'{self.URL}/video/{path}',
+                "headers": {
+                    "Authorization": f"bearer {self.token}",
+                },
+            }
+            response = requests.request(**request)
+            if response.status_code == 200:
+                pp = p.parent
+                if not pp.exists():
+                    pp.mkdir(parents=True, exist_ok=True)
+                p.write_bytes(response.content)
+            logger.info('Video file %s finished downloading', path)
+        else:
+            logger.info('Video file %s already exists', path)
 
 
     async def get_videos_metadata_paginated(self, environment_id, start_date, end_date, camera_id=None, skip=0, limit=100):
@@ -161,7 +160,7 @@ class VideoStorageClient:
     async def get_videos_metadata(self, environment_id, start_date, end_date, camera_id=None, skip=0, limit=100):
         request = {
             "method": "GET",
-            "url": f'{self.DOMAIN}/videos/{environment_id}/device/{camera_id}' if camera_id is not None else f'{self.DOMAIN}/videos/{environment_id}',
+            "url": f'{self.URL}/videos/{environment_id}/device/{camera_id}' if camera_id is not None else f'{self.URL}/videos/{environment_id}',
             "headers": {
                 "Authorization": f"bearer {self.token}",
             },
@@ -180,12 +179,13 @@ class VideoStorageClient:
 
 
     async def upload_video(self, path, local_cache_directory=None):
-        prefix = local_cache_directory if local_cache_directory is not None else VIDEO_STORAGE_LOCAL_CACHE_DIRECTORY
-        full_path = os.path.join(prefix, path)
+        if local_cache_directory is None:
+            local_cache_directory=self.CACHE_DIRECTORY
+        full_path = os.path.join(local_cache_directory, path)
         ptype, file_details = parse_path(path)
         if ptype == "file":
             file_details["path"] = full_path
-            file_details["filepath"] = full_path[len(prefix):]
+            file_details["filepath"] = full_path[len(local_cache_directory):]
             resp =  await self.upload_videos([file_details])
             return resp[0]
         return {"error": "invalid path. doesn't match pattern [environment_id]/[camera_id]/[year]/[month]/[day]/[hour]/[min]-[second].mp4"}
@@ -194,7 +194,7 @@ class VideoStorageClient:
     async def upload_videos(self, file_details: List[Dict]):
         request = {
             "method": "POST",
-            "url": f"{self.DOMAIN}/videos",
+            "url": f"{self.URL}/videos",
             "headers": {
                 "Authorization": f"bearer {self.token}",
             }
@@ -231,7 +231,7 @@ class VideoStorageClient:
     async def video_existence_check(self, paths: List[str]):
         request = {
             "method": "POST",
-            "url": f"{self.DOMAIN}/videos/check",
+            "url": f"{self.URL}/videos/check",
             "headers": {
                 "Authorization": f"bearer {self.token}",
             },
@@ -247,7 +247,15 @@ class VideoStorageClient:
             return [{"err": "response error", "path": p, "exists": False} for p in paths]
 
 
-    async def upload_videos_in(self, path, local_cache_directory=None, batch_size=4):
+    async def upload_videos_in(
+        self,
+        path,
+        local_cache_directory=None,
+        batch_size=video_io.config.SYNC_BATCH_SIZE,
+        max_workers=video_io.config.MAX_SYNC_WORKERS
+    ):
+        if local_cache_directory is None:
+            video_io = self.CACHE_DIRECTORY
         t, details = parse_path(path[:-1] if path[-1] == '/' else path)
         if details:
             if t == "file":
@@ -257,19 +265,18 @@ class VideoStorageClient:
             if t == "month":
                 raise SyncError("cannot sync a month of videos, try limiting to a day")
             files_found = []
-            prefix = local_cache_directory if local_cache_directory is not None else VIDEO_STORAGE_LOCAL_CACHE_DIRECTORY
-            for root, _, files in os.walk(os.path.join(prefix, path)):
+            for root, _, files in os.walk(os.path.join(local_cache_directory, path)):
                 for file in files:
                     full_path = os.path.join(root, file)
-                    ptype, file_details = parse_path(full_path[len(prefix):])
+                    ptype, file_details = parse_path(full_path[len(local_cache_directory):])
                     if ptype == "file":
                         file_details["path"] = full_path
-                        file_details["filepath"] = full_path[len(prefix):]
+                        file_details["filepath"] = full_path[len(local_cache_directory):]
                         files_found.append(file_details)
             details["files_found"] = len(files_found)
             details["files_uploaded"] = 0
             details["details"] = []
-            with concurrent.futures.ThreadPoolExecutor(max_workers=MAX_SYNC_WORKERS) as executor:
+            with concurrent.futures.ThreadPoolExecutor(max_workers=max_workers) as executor:
                 results = executor.map(self.upload_videos, chunks(files_found, batch_size))
                 for future in results:
                     result = await future
